fix #18525: the Instrument category has the highest priority for compatibility reasons
