+++ src/framework/audio/internal/plugins/knownaudiopluginsregister.cpp
@@ -22,6 +22,7 @@
 
 #include "knownaudiopluginsregister.h"
 
+#include "audioutils.h"
 #include "serialization/json.h"
 
 #include "log.h"
